requirements.txt
# Core dependencies
pyspark==3.5.0
databricks-sdk==0.18.0
databricks-cli==0.18.0

# ML and data processing
mlflow==2.9.2
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.3
scipy==1.11.4

# Deep Learning
tensorflow==2.15.0
keras==2.15.0

# Data validation
great-expectations==0.18.8
pydantic==2.5.3

# Feature engineering
category-encoders==2.6.3
imbalanced-learn==0.11.0

# Monitoring and alerting
prometheus-client==0.19.0
requests==2.31.0

# Utilities
PyYAML==6.0.1
python-dotenv==1.0.0
pytz==2023.3
